src/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
boto3==1.20.53
botocore==1.23.53
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.10.8
char-featurizer==1.0.0
charset-normalizer==2.0.10
click==8.0.3
configparser==5.2.0
datasets==1.10.0
dill==0.3.4
docker-pycreds==0.4.0
docopt==0.6.2
et-xmlfile==1.1.0
FastNLP==0.7.0
filelock==3.4.2
fitlog==0.9.13
Flask==2.1.2
flatbuffers==1.12
frozenlist==1.3.0
fsspec==2022.2.0
gast==0.4.0
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.25
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
mock==4.0.3
multidict==6.0.2
multiprocess==0.70.12.2
networkx==2.6.3
nltk==3.7
numpy==1.21.5
nvidia-htop==1.0.5
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
prettytable==3.3.0
promise==2.3
protobuf==3.19.1
psutil==5.9.0
pyahocorasick==1.4.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pypinyin==0.46.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-crf==0.7.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rouge==1.0.1
rsa==4.8
s3transfer==0.5.1
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.1
seqeval==1.2.2
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
thulac==0.2.1
tokenizers==0.12.1
torch==1.9.0+cu111
tqdm==4.62.3
transformers==4.19.4
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.9
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
xxhash==3.0.0
yarl==1.7.2
yaspin==2.1.0
zhconv==1.4.3
zhon==1.1.5
zipp==3.7.0
